Make the gtk+ portions of the project optional.
This commit first separates gpicker from gpicker-simple, making them
fully separate binaries, built with different flags. This makes it
easy to then make the building of gpicker optional.

On systems on which gtk+ or gthread is unavailable, or if configure is
passed the argument --without-gtk, only gpicker-simple is available.
gpicker isn't built at all.